V14 ItemAttackApplicationV2 Region AOEs

+++ module/actor/actor.mjs
@@ -10,7 +10,7 @@ import {
     tokenEducatedGuess,
 } from "../utility/util.mjs";
 import { HeroProgressBar } from "../utility/progress-bar.mjs";
-import { clamp } from "../utility/compatibility.mjs";
+import { clamp, isGameV14OrLater } from "../utility/compatibility.mjs";
 import { overrideCanAct } from "../settings/settings-helpers.mjs";
 import { roundFavorPlayerTowardsZero, roundFavorPlayerAwayFromZero } from "../utility/round.mjs";
 import { HeroItemCharacteristic } from "../item/HeroSystem6eTypeDataModels.mjs";
@@ -2443,14 +2443,19 @@ export class HeroSystem6eActor extends HeroObjectCacheMixin(Actor) {
                 // Delete maneuvers (or any other existing items) that don't
                 // match template prior to possibly changing is5e
                 if (heroJson.CHARACTER.TEMPLATE !== this.system.CHARACTER?.TEMPLATE) {
-                    const itemsToDeleteIs5e = this.items
-                        .filter((i) => i.system.is5e !== this.is5ePreview(heroJson.CHARACTER.TEMPLATE))
-                        .map((m) => m.id);
+                    const itemsToDeleteIs5e = this.items.filter(
+                        (i) => i.system.is5e !== this.is5ePreview(heroJson.CHARACTER.TEMPLATE),
+                    );
+
                     if (itemsToDeleteIs5e.length > 0) {
                         console.warn(`Deleting ${itemsToDeleteIs5e.length} is5e mismatches`);
-                        await this.deleteEmbeddedDocuments("Item", itemsToDeleteIs5e, {
-                            render: false,
-                        });
+                        if (itemsToDeleteIs5e.length === this.items.size) {
+                            await this.items.clear();
+                        } else {
+                            await this.deleteEmbeddedDocuments("Item", itemsToDeleteIs5e, {
+                                render: false,
+                            });
+                        }
                     }
                 }
 
@@ -2494,8 +2499,23 @@ export class HeroSystem6eActor extends HeroObjectCacheMixin(Actor) {
                     .replace("competentpc", "pc"); // super old HDC
 
                 if (targetType && this.type.replace("npc", "pc") !== targetType) {
-                    if (Object.keys(game.system.template.Actor).includes(targetType)) {
-                        await this.update({ type: targetType, [`==system`]: this.system });
+                    if (Object.keys(game.system.documentTypes.Actor).includes(targetType)) {
+                        if (isGameV14OrLater()) {
+                            // REF: https://github.com/foundryvtt/foundryvtt/issues/13090
+                            // AARON WAS HERE on 4/4/2026: Update fails, likely a foundry bug.
+                            // Error: The type of a Document may only be changed if the system field
+                            //        is also updated with a ForcedReplacement operator.
+                            // A subsequent upload works, not ready for publish.
+                            await this.update(
+                                {
+                                    type: targetType,
+                                    system: this.system,
+                                },
+                                { recursive: false },
+                            );
+                        } else {
+                            await this.update({ type: targetType, [`==system`]: this.system });
+                        }
                     } else {
                         ui.notifications.error(`${targetType} is not a valid actor type`);
                     }
@@ -2546,11 +2566,11 @@ export class HeroSystem6eActor extends HeroObjectCacheMixin(Actor) {
 
             // NOTE don't put this into the promiseArray because we create things in here that are absolutely required by later items (e.g. strength placeholder).
             // if (this.type === "pc" || this.type === "npc" || this.type === "automaton") {
-            uploadProgressBar.advance(`${this.name}: Evaluating non HDC items for PCs, NPCs, and Automatons`, 0);
+            uploadProgressBar.advance(`${this.name}: addFreeStuff`, 0);
 
             await this.addFreeStuff();
 
-            uploadProgressBar.advance(`${this.name}: Evaluated non HDC items for PCs, NPCs, and Automatons`, 0);
+            uploadProgressBar.advance(`${this.name}: addFreeStuff completed`, 0);
             //}
 
             uploadPerformance.progressBarFreeStuff - uploadPerformance._d;
